+++ test/mocks/MockERC6551Account.sol
@@ -20,7 +20,7 @@ contract MockERC6551Account is IERC165, IERC6551Account, IERC6551Executable {
         _initialized = val;
     }
 
-    function execute(address, uint256, bytes calldata, uint256)
+    function execute(address, uint256, bytes calldata, uint8)
         external
         payable
         returns (bytes memory)
